Sync with gnulib f345edc03da1a81f313dd70305764025bffeb1fa.
Port 'open' and 'close' redefinitions to AIX 7.1.

Do not #undef 'open' and 'close'. AIX 7 does '#define open open64' and
then 'int open64(const char *, int, ...);', which means the declaration
for 'open' gets lost if we later '#undef open'. Discovered while
building grep pretest 2.18.151-1c770 on AIX 7.1, where the compilation
reported the non-fatal error "In function 'openat_proc_name' ...
warning: implicit declaration of function 'open'". In this case the
error is relatively harmless, but in other cases it might not be so
minor.

Signed-off-by: Andrew Borodin <aborodin@vmail.ru>